Group all build extended tests into one file
